Framework localization fix
Previously QBValidator was using `NSLocalizedStringFromTable` to do localizations.   Which works in the case that you're not using QBValidator as a .framework.

The solution is to first try to `NSLocalizedStringFromTableInBundle` using `[NSBundle bundleForClass:[self class]]`.  This will get the key from the .framework bundle if available.   If not, then go to the main bundle