Copy min/max Value properties
Previously the minValue and maxValue properties on QBValidationRangeRule were set to assign.   Which meant that whoever created this rule was in charge of keeping reference to these objects until the rule finished running.

This practice is not consistent with the interface designed by QBValidator.   The use case is defined to use the macro functions as such

```
QBVLessThan(@(10))
QBVInRange(10, 20)
```

In the examples above, nobody is keeping hold of the `@(10)`.   So we basically enter into a race condition.   The device could reassign that pointer without our knowledge and cause a crash whenever the comparisons are made.

The fix is to simply copy these values, so the rule has a dedicated reference to the pointers that will be around for the lifetime of the rule.